Modularize services
